Read LOG_LEVEL env var or default to :debug
This enables to define the log level on demand. Currently, debug is
generating high volume of logs and only ~30min fit into Heroku's limits.

We went with Coralogix as log management add-on (they have a trial plus
freemium) but we want to have this ENV var as a fallback. Reducing the
volume will let us fit more into the 1500-lines limit.