Move rack-timeout to Bundler production group
It's very annoying in development because if you don't specify the
RACK_TIMEOUT_SERVICE_TIMEOUT env var it defaults to 15s of response
time, and that's very common to hit when booting the server.

I also took the chance to reorder gems within the production group so
Rubocop shuts up.