Point out our own stable 0.22 with extra fixes
Using the same branch for both suggesting a fix to upstream and pointing
to it in production is becoming a dangerous mess. I'm doing/undoing
changes and rewriting the git history but I don't want to risk
introducing regressions in production.

Once that patch is merged we can consider dropping our stable branch.
Meanwhile, we better off keep things decoupled.